Add `FullMerkleTree` struct for testing
The main `MerkleTree` implementation is not storing all inodes, but it
rather stores subtree hashes - which makes sense, because we support
only inserting new values and that way we save storage.

But to test whether `MerkleTree::insert` works properly, we need to
check it with some simplier implementation. This is the purpose of
`FullMerkleTree`, which contains all inodes and has simplier `insert`
implementation.