Add `Hasher` trait and its implementations
We used to have a separate repo for that trait, but it's unlikely to be
used anywhere outside the Merkle tree context, hence moving it here.

Starting from now, `anchor-lang` becomes a non-optional dependency of
`light-merkle-tree`, **but** the `solana` feature still determines
whether the `MerkleTree` struct implements `AnchorSerialize` and all the
other traits needed for storing it in a Solana account.

Without `solana` feature, this crate is still not going to be usable for
Solana programs, but rather only as a regular Rust crate to be used on
host environments (mainly for testing). The only purpose of dependency
on `anchor-lang` in that case is providing hash functions.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>